Fix a couple more tests
