+++ .travis.yml
@@ -7,6 +7,7 @@ android:
     - build-tools-25.0.2
     - extra-android-m2repository
     - android-25
+    - sys-img-armeabi-v7a-android-25
   licenses:
     - 'android-sdk-license-.+'
 
@@ -14,7 +15,7 @@ jdk:
   - oraclejdk8
 
 before_script:
-  - echo no | android create avd --force -n test -t android-22 --abi armeabi-v7a
+  - echo no | android create avd --force -n test -t android-25 --abi armeabi-v7a
   - emulator -avd test -no-audio -no-window &
   - android-wait-for-emulator
   - adb shell input keyevent 82 &
