+++ betamax-tests/src/test/groovy/software/betamax/AnnotationTest.groovy
@@ -26,10 +26,8 @@ import org.junit.BeforeClass
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.junit.runners.BlockJUnit4ClassRunner
-import org.junit.runners.model.FrameworkMethod
-import software.betamax.junit.Betamax
 import software.betamax.android.RecorderRule
+import software.betamax.junit.Betamax
 import software.betamax.proxy.BetamaxInterceptor
 import spock.lang.Shared
 
@@ -114,32 +112,5 @@ class AnnotationTest {
 
     assert response.code() == HTTP_OK
     assert response.header(VIA) == null
-  }
-
-}
-
-/**
- * This is an evil hack. JUnit does not guarantee test execution order and the methods in this test depend on each
- * other. In particular `annotatedTestCanPlayBack` will fail if run before `annotatedTestCanRecord`. Really the tests
- * should be idempotent.
- */
-class OrderedRunner extends BlockJUnit4ClassRunner {
-  private static final ORDER = [
-      'noTapeIsInsertedIfThereIsNoAnnotationOnTheTest',
-      'annotationOnTestCausesTapeToBeInserted',
-      'tapeIsEjectedAfterAnnotatedTestCompletes',
-      'annotatedTestCanRecord',
-      'annotatedTestCanPlayBack',
-      'canMakeUnproxiedRequestAfterUsingAnnotation'
-  ]
-
-  OrderedRunner(Class testClass) {
-    super(testClass)
-  }
-
-  @Override protected List<FrameworkMethod> computeTestMethods() {
-    super.computeTestMethods().sort(false) { FrameworkMethod o1, FrameworkMethod o2 ->
-      ORDER.indexOf(o1.name) <=> ORDER.indexOf(o2.name)
-    }
   }
 }
